% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_insee_metadata.R
\name{add_insee_metadata}
\alias{add_insee_metadata}
\title{Add metadata to the raw data}
\usage{
add_insee_metadata(df, label = TRUE)
}
\arguments{
\item{df}{a dataframe containing data obtained from get_insee_idbank or get_insee_dataset}

\item{label}{if TRUE, labels for the dimensions are provided}
}
\value{
a tibble with the data given as parameter plus the corresponding metadata
}
\description{
Add metadata to the raw data
}
\details{
Add metadata to the raw data obtained from get_insee_idbank or get_insee_dataset
}
\examples{
\donttest{
library(tidyverse)

data =
 get_insee_idbank("001694061") \%>\%
 add_insee_metadata()
}
}
