% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_title.R
\name{get_column_title}
\alias{get_column_title}
\title{Get the title of dataset's columns}
\usage{
get_column_title(dataset = NULL)
}
\arguments{
\item{dataset}{an INSEE's dataset, if NULL}
}
\value{
a dataframe
}
\description{
Get the title of dataset's columns
}
\examples{
\donttest{
column_titles = get_column_title("CNA-2014-CONSO-MEN")
}
}
