% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{get_idbank_list}
\alias{get_idbank_list}
\title{Download a mapping dataset betwen INSEE series keys (idbank) and SDMX series names}
\usage{
get_idbank_list(
  dataset = NULL,
  insee_path = "https://www.insee.fr/en/statistiques",
  mapping_file_path = "/fichier/2868055/correspondance_idbank_dimension.zip",
  mapping_file_pattern = "correspondance_idbank_dimension",
  mapping_file_sep = ";",
  idbank_nchar = 9
)
}
\arguments{
\item{dataset}{if a dataset name is provided, only a subset of the data is delivered, otherwise
all the data is returned}

\item{insee_path}{main INSEE link}

\item{mapping_file_path}{INSEE zip file containting idbank dataset ()}

\item{mapping_file_pattern}{dataset file name pattern}

\item{mapping_file_sep}{dataset file separator}

\item{idbank_nchar}{number of character in an series key (idbank)}
}
\description{
Download a mapping dataset betwen INSEE series keys (idbank) and SDMX series names
}
\details{
Download a mapping dataset betwen INSEE series keys (idbank) and SDMX series names
}
\examples{
idbank_all = get_idbank_list()
idbank_ipc = get_idbank_list("IPC-2015")

}
