% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_insee}
\alias{get_insee}
\title{Get data from INSEE BDM database with a SDMX query link}
\usage{
get_insee(link)
}
\arguments{
\item{link}{SDMX query link}
}
\description{
Get data from INSEE BDM database with a SDMX query link
}
\details{
Get data from INSEE BDM database with a SDMX query link
Mainly for package internal use from the functions get_insee_dataset and get_insee_idbank
}
\examples{
insee_link = "http://www.bdm.insee.fr/series/sdmx/data/SERIES_BDM"
insee_query = file.path(insee_link, paste0("010539365","?", "firstNObservations=1"))
data = get_insee(insee_query)

}
