% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_insee_idbank}
\alias{get_insee_idbank}
\title{Get data from INSEE series idbank}
\usage{
get_insee_idbank(
  ...,
  startPeriod = NULL,
  endPeriod = NULL,
  firstNObservations = NULL,
  lastNObservations = NULL
)
}
\arguments{
\item{...}{one or several series key (idbank)}

\item{startPeriod}{start date of data}

\item{endPeriod}{end date of data}

\item{firstNObservations}{get the first N observations for each key series (idbank)}

\item{lastNObservations}{get the last N observations for each key series (idbank)}
}
\description{
Get data from INSEE series idbank
}
\details{
Get data from INSEE series idbank
}
\examples{
\donttest{

#example 1 : import price index of industrial products and turnover index : manufacture of wood
data = get_insee_idbank("001558315", "010540726")

#example 2 : unemployment data

library(tidyverse)

idbank_list = get_idbank_list()

df_idbank_list_selected =
  idbank_list \%>\%
  filter(nomflow == "CHOMAGE-TRIM-NATIONAL") \%>\%  #unemployment dataset
  filter(dim5 == 0) \%>\% #men and women
  mutate(title = get_insee_title(idbank))

idbank_list_selected = df_idbank_list_selected \%>\% pull(idbank)

unem = get_insee_idbank(idbank_list_selected)
}

}
