% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_insee_title.R
\name{get_insee_title}
\alias{get_insee_title}
\title{Get title from INSEE series idbank}
\usage{
get_insee_title(..., lang = "en")
}
\arguments{
\item{...}{list of series key (idbank)}

\item{lang}{language of the title, by default it is Engligh, if lang is different from "en" then French will be the title's language}
}
\value{
a character vector with the titles
}
\description{
Get title from INSEE series idbank
}
\details{
Query INSEE website to get series title from series key (idbank).
Any query to INSEE database can handle around 400 idbanks at maximum, if necessary the idbank list will then be splitted in several lists of 400 idbanks each.
Consequently, it is not advised to use it on the whole idbank dataset, the user should filter the idbank dataset first.
}
\examples{
\donttest{
#example 1 : industrial production index on manufacturing and industrial activities
title = get_insee_title("010537900")

#example 2 : automotive industry and overall industrial production
library(tidyverse)

idbank_list = get_idbank_list()

idbank_list_selected =
  idbank_list \%>\%
  filter(nomflow == "IPI-2015") \%>\% #industrial production index dataset
  filter(dim1 == "M") \%>\% #monthly
  filter(dim6 == "INDICE") \%>\% #index
  filter(dim9  == "CVS-CJO") \%>\% #Working day and seasonally adjusted SA-WDA
  filter(str_detect(dim5,"^29$|A10-BE")) \%>\% #automotive industry and overall industrial production
  mutate(title = get_insee_title(idbank))
}

}
