% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_release.R
\name{get_last_release}
\alias{get_last_release}
\title{Get the datasets released in the last 30 days}
\usage{
get_last_release()
}
\value{
a tibble with the data
}
\description{
Get the datasets released in the last 30 days
}
\examples{
\donttest{
last_release = get_last_release()
}

}
