% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{get_dataset_list}
\alias{get_dataset_list}
\title{Download an INSEE dataset list}
\usage{
get_dataset_list(link_dataflow = "https://bdm.insee.fr/series/sdmx/dataflow")
}
\arguments{
\item{link_dataflow}{SDMX query link}
}
\description{
Download an INSEE dataset list
}
\details{
the datasets returned are the ones available through a SDMX query
}
\examples{
dataset = get_dataset_list()

}
