% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{get_idbank_list}
\alias{get_idbank_list}
\title{Download a mapping dataset betwen INSEE series keys (idbank) and SDMX series names}
\usage{
get_idbank_list(dataset = NULL)
}
\arguments{
\item{dataset}{if a dataset name is provided, only a subset of the data is delivered, otherwise
all the data is returned}
}
\description{
Download a mapping dataset betwen INSEE series keys (idbank) and SDMX series names
}
\details{
Download a mapping dataset betwen INSEE series keys (idbank) and SDMX series names
}
