% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_insee}
\alias{get_insee}
\title{Get data from INSEE BDM database with a SDMX query link}
\usage{
get_insee(link)
}
\arguments{
\item{link}{SDMX query link}
}
\description{
Get data from INSEE BDM database with a SDMX query link
}
\details{
Get data from INSEE BDM database with a SDMX query link
Mainly for package internal use from the functions get_insee_dataset and get_insee_idbank
}
\examples{
data = get_insee("http://www.bdm.insee.fr/series/sdmx/data/SERIES_BDM/010539365")

}
