% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_insee_idbank}
\alias{get_insee_idbank}
\title{Get data from INSEE series idbank}
\usage{
get_insee_idbank(
  ...,
  startPeriod = NULL,
  endPeriod = NULL,
  firstNObservations = NULL,
  lastNObservations = NULL,
  insee_bdm_series_link = "http://www.bdm.insee.fr/series/sdmx/data/SERIES_BDM"
)
}
\arguments{
\item{...}{one or several series key (idbank)}

\item{startPeriod}{start date of data}

\item{endPeriod}{end date of data}

\item{firstNObservations}{get the first N observations for each key series (idbank)}

\item{lastNObservations}{get the last N observations for each key series (idbank)}

\item{insee_bdm_series_link}{SDMX link used to build the query on series keys}
}
\description{
Get data from INSEE series idbank
}
\details{
Get data from INSEE series idbank
}
\examples{
data = get_insee_idbank("001558315", "010540726")

}
