% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{search_insee}
\alias{search_insee}
\title{Search among insee dataset names a pattern}
\usage{
search_insee(pattern)
}
\arguments{
\item{pattern}{string used to filtern the dataset list}
}
\description{
Search among insee dataset names a pattern
}
\details{
the function ignores accents and cases
}
\examples{
\donttest{
# example 1 : search one pattern
#the correct word in French would be 'enqu\U00EAte' which means survey
dataset_enquete = search_insee("enquete")

# example 2 : search multiple patterns
dataset_survey_gdp = search_insee("Survey|gdp")
}
}
